feat: add refresh on focus to feed
