+++ src/screens/MainScreen/MainScreen.js
@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { View } from 'react-native';
+import { useFocusEffect } from '@react-navigation/native';
+import { useDispatch } from 'react-redux';
 
+import { getFeed } from 'actions/feedActions';
 import { MAIN_SCREEN } from 'constants/screens';
 import useNotifications from 'hooks/useNotifications';
 
@@ -10,9 +13,20 @@ import styles from './MainScreen.styles';
 
 const MainScreen = () => {
   useNotifications();
+
+  const dispatch = useDispatch();
+
+  const handleRefresh = useCallback(() => {
+    dispatch(getFeed({ shouldReplace: true }));
+  }, [dispatch]);
+
+  useFocusEffect(() => {
+    handleRefresh();
+  }, [handleRefresh]);
+
   return (
     <View style={styles.container} testID={MAIN_SCREEN}>
-      <FeedList />
+      <FeedList handleRefresh={handleRefresh} />
     </View>
   );
 };
